+++ apworld/Rules.py
@@ -3693,6 +3693,7 @@ class UltrakillRules:
                             or shoalt_any(state)
                             or rock_any(state)
                         )
+                        and arm2(state)
                         and skull(state, "8-3", "Blue")
                     )
                     or (
@@ -4645,6 +4646,7 @@ class UltrakillRules:
                                     or shoalt_any(state)
                                     or rock_any(state)
                                 )
+                                and arm2(state)
                                 and skull(state, "8-3", "Blue")
                             )
                         )